Refactor symbol parsing logic to a new function `parse_and_get_error_symbols` for DRY compliance.
